+++ Sources/MapboxMaps/Foundation/Extensions/Core/MapSnapshotOptions.swift
@@ -10,6 +10,10 @@ public struct MapSnapshotOptions: Sendable {
     /// By default, `GlyphsRasterizationOptions` will use `.ideographsRasterizedLocally`
     public var glyphsRasterizationOptions: GlyphsRasterizationOptions
 
+    /// Scale factor to scale map shapshot icons and texts. Default is `1.0`.
+    @_spi(Experimental)
+    public var scaleFactor: CGFloat
+
     /// Flag that determines if the logo should be shown on the snapshot
     public var showsLogo: Bool
 
@@ -25,12 +29,15 @@ public struct MapSnapshotOptions: Sendable {
     ///         `.ideographsRasterizedLocally`
     ///   - showsLogo: Flag that determines if the logo should be shown on the snapshot
     ///   - showsAttribution: Flag that determines if attribution should be shown on the snapshot
+    ///   - scaleFactor: Scale factor to scale map snapshot icons and texts. Default is `1.0`.
     public init(size: CGSize,
                 pixelRatio: CGFloat,
                 glyphsRasterizationOptions: GlyphsRasterizationOptions = GlyphsRasterizationOptions(),
                 showsLogo: Bool = true,
-                showsAttribution: Bool = true) {
+                showsAttribution: Bool = true,
+                scaleFactor: CGFloat = 1.0) {
         precondition(pixelRatio > 0)
+        precondition(scaleFactor > 0)
         precondition(size.width * pixelRatio <= 8192, "Width or scale too great.")
         precondition(size.height * pixelRatio <= 8192, "Height or scale too great.")
 
@@ -39,6 +46,7 @@ public struct MapSnapshotOptions: Sendable {
         self.glyphsRasterizationOptions = glyphsRasterizationOptions
         self.showsLogo = showsLogo
         self.showsAttribution = showsAttribution
+        self.scaleFactor = scaleFactor
     }
 }
 
@@ -46,7 +54,8 @@ extension MapSnapshotOptions {
     internal init(_ objcValue: CoreMapSnapshotOptions) {
         self.init(size: CGSize(objcValue.__size),
                   pixelRatio: CGFloat(objcValue.pixelRatio),
-                  glyphsRasterizationOptions: objcValue.glyphsRasterizationOptions ?? GlyphsRasterizationOptions())
+                  glyphsRasterizationOptions: objcValue.glyphsRasterizationOptions ?? GlyphsRasterizationOptions(),
+                  scaleFactor: CGFloat(objcValue.__scaleFactor))
     }
 }
 
@@ -56,6 +65,7 @@ extension CoreMapSnapshotOptions {
 
         self.init(__size: Size(width: Float(size.width), height: Float(size.height)),
                   pixelRatio: Float(swiftValue.pixelRatio),
-                  glyphsRasterizationOptions: swiftValue.glyphsRasterizationOptions)
+                  glyphsRasterizationOptions: swiftValue.glyphsRasterizationOptions,
+                  scaleFactor: Float(swiftValue.scaleFactor))
     }
 }