+++ Tests/MapboxMapsTests/Foundation/MapboxMapTests.swift
@@ -56,7 +56,8 @@ final class MapboxMapTests: XCTestCase {
             crossSourceCollisions: mapInitOptions.mapOptions.crossSourceCollisions.NSNumber,
             size: mapInitOptions.mapOptions.size.map(Size.init),
             pixelRatio: mapInitOptions.mapOptions.pixelRatio,
-            glyphsRasterizationOptions: nil) // __map.getOptions() always returns nil for glyphsRasterizationOptions
+            glyphsRasterizationOptions: nil, // __map.getOptions() always returns nil for glyphsRasterizationOptions
+            scaleFactor: 1.0)
 
         let actualMapOptions = mapboxMap.options